Font Fallback
You can now set multiple fonts that can be used for fallback purposes.
Each font and its options are delimited by a comma, and there must not
be spaces before/after the comma. As an example,
"set guifont=Consolas,Meiryo\ UI", will set "Consolas" as the primary
font, and "Meiryo UI" as a fallback font. This (should) works with any
number of fonts.