Add commands to edit title bar colors
The default behavior of the title bar is to change color according to
Neovim's default colors.
The command :NvuiTitlebarColors (fg) (bg) changes the titlebar's
foreground and background colors, and gets rid of the default behavior.
For example, :NvuiTitlebarColors #000 #fff will set the foreground color
to black, and the background color to white, and the title bar color
will no longer change with the default colors.
You can go back to the default behavior using :NvuiTitlebarUnsetColors.