Add cursor animations
Looks pretty nice. No fancy effects like Neovide but the cursor's
movement is animated. Frametimes / framerate is controllable with
:NvuiCursorFrametime {ms}, duration controllable with
:NvuiCursorAnimationDuration {seconds}, and the scaler is controllable
with :NvuiCursorScaler {scaler}. Also added the documentation for these
commands.
This has also alerted me to deficiencies in the move and scroll
animations, particularly the scroll animation, that is caused by the
timer reset.