[WIP] Various Changes (#68)
* Move some templates to header, using template in send_response

Also in main, stop looking for arguments when we reach the -- flag.

* Remove msgpackc-cxx linking

* Use Cmd for C- key on Mac, handle Meta (D-) key

* Use floating point font metrics

* Return ptr to placed value in LRU cache

This prevents a second lookup to get the stored value

* Cleanup

Remove Nvim::send_request_sync, Nvim::eval, Nvim::get_api_info
and clean up Nvim::read_output

* Clean up debug output

* Link cmdline to default colors by default

* Bounds checks in grid_line, (potential) fix for macOS input

* Search grids for mouse input in order

* Fix cursor shape on non-frameless window

* Hide info widget, fix alignment of popupmenu info

Only works with multigrid enabled since this is a workaround around a
bug with Neovim. Info widget is hidden since Neovim still seems to be
showing the info as a grid.

* Print error if font family not found

* Fix pum info positioning, use Neovim-Qt's convertKey to process key

presses. Popup menu info positioning is still a little broken since if
you have a signature preview open and the popup menu opens the preview
gets attached to the popup menu which is not good.

* Add includes

* Use floating font metrics everywhere, don't load new icons after

program starts

* Swap around D- and C- modifiers on Mac

This should make Ctrl map to <C->, Cmd map to <D-> which is the opposite
of how it was before.