docs: add instruction for Arch Linux users
The Ubuntu instruction is not useful for Arch Linux user. And the AUR
package is now still unusable. So I think it is necessary to have
document for user who use Arch Linux to know how to build.

This instruction include: what dependencies is needed, how to clone, how
to build and how to install.

Signed-off-by: Avimitin <avimitin@gmail.com>