Separate text and background drawing, clip cursor text
On Windows Direct2D is being a pain by not filling rectangles properly.
The problem is that when we are drawing the background rect, the x and y
values are floating points, so sometimes the actual pixel boundaries
don't line up and so there ends up being an undrawn area of the bitmap
which looks weird. The first answer was to floor the starting x-position
and ceil the ending x-position in draw_text_and_bg but that has the
weird effect of making text shift sometimes (i.e. when you select it)
because contiguous pieces of text are drawn together, but when you break
the contiguousness (by having a different highlight for the text) it
suddenly moves. So the answer is to only use floor for the background
drawing, while keeping text drawing the same.